+++ e2e/node/smoke.e2e.ts
@@ -1,5 +1,7 @@
+import assert from "node:assert";
 import { execSync } from "node:child_process";
-import { mkdir, rm, writeFile } from "node:fs/promises";
+import { mkdir, readFile, rm, writeFile } from "node:fs/promises";
+import { createRequire } from "node:module";
 import { resolve } from "node:path";
 import { test } from "node:test";
 import { testRootDir } from "./utils.js";
@@ -14,17 +16,17 @@ const { Document, MetadataMode, VectorStoreIndex } = require('llamaindex')
 const { OpenAIEmbedding } = require('@llamaindex/openai')
 const { Settings } = require('@llamaindex/core/global')`;
   const mainCode = `
-async function main() {
-  Settings.embedModel = new OpenAIEmbedding({
-    model: 'text-embedding-3-small',
-    apiKey: '${process.env.OPENAI_API_KEY}',
-  })
-  const model = Settings.embedModel
-  if (model == null) {
-    process.exit(-1)
-  }
+Settings.embedModel = new OpenAIEmbedding({
+  model: 'text-embedding-3-small',
+  apiKey: '${process.env.OPENAI_API_KEY}',
+})
+const model = Settings.embedModel
+if (model == null) {
+  process.exit(-1)
 }
-main().catch(console.error)`;
+const document = new Document({ text: 'Hello, world!' })
+const index = await VectorStoreIndex.fromDocuments([document])
+`;
   t.before(async () => {
     await mkdir(resolve(testRootDir, ".temp"), {
       recursive: true,
@@ -83,4 +85,21 @@ main().catch(console.error)`;
       });
     });
   }
+});
+
+test('no extra deps in "@llamaindex/env" cjs module', async () => {
+  const modules = ["@aws-crypto/sha256-js"];
+  const require = createRequire(import.meta.url);
+  const envPackage = require.resolve("@llamaindex/env");
+  const file = await readFile(envPackage, "utf-8");
+  for (const module of modules) {
+    assert.ok(!file.includes(module));
+  }
+});
+
+test('no error when require "llamaindex" in CJS', async () => {
+  const code = `require('llamaindex')`;
+  execSync(`${process.argv[0]} -e "${code}"`, {
+    cwd: process.cwd(),
+  });
 });