+++ packages/llamaindex/src/index.ts
@@ -1,21 +1,9 @@
 export * from "./index.edge.js";
-export * from "./readers/index.js";
-export * from "./storage/index.js";
-// Exports modules that doesn't support non-node.js runtime
-export {
-  HuggingFaceEmbedding,
-  HuggingFaceEmbeddingModelType,
-} from "./embeddings/HuggingFaceEmbedding.js";
 
-export {
-  GeminiVertexSession,
-  type VertexGeminiSessionOptions,
-} from "@llamaindex/google";
-
-// Expose AzureDynamicSessionTool for node.js runtime only
-export { AzureDynamicSessionTool } from "@llamaindex/azure";
+// TODO: clean up, move to jinaai package
 export { JinaAIEmbedding } from "./embeddings/JinaAIEmbedding.js";
 
-// Don't export vector store modules for non-node.js runtime on top level,
+// Don't export file-system stores for non-node.js runtime on top level,
 //  as we cannot guarantee that they will work in other environments
+export * from "./storage/index.js";
 export * from "./vector-store.js";