+++ e2e/examples/llama-parse-browser/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @llamaindex/llama-parse-browser-test
 
+## 0.0.46
+
+### Patch Changes
+
+- @llamaindex/cloud@3.0.1
+
+## 0.0.45
+
+### Patch Changes
+
+- @llamaindex/cloud@3.0.0
+
 ## 0.0.44
 
 ### Patch Changes