+++ packages/autotool/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @llamaindex/autotool
 
+## 6.0.1
+
+### Patch Changes
+
+- Updated dependencies [6d37d44]
+  - llamaindex@0.9.1
+
+## 6.0.0
+
+### Patch Changes
+
+- Updated dependencies [6a4a737]
+- Updated dependencies [d924c63]
+- Updated dependencies [b490376]
+- Updated dependencies [f4588bc]
+  - llamaindex@0.9.0
+
 ## 5.0.37
 
 ### Patch Changes