+++ packages/providers/anthropic/CHANGELOG.md
@@ -1,5 +1,31 @@
 # @llamaindex/anthropic
 
+## 0.1.2
+
+### Patch Changes
+
+- be74207: fix: dont add empty text block to tool call
+
+## 0.1.1
+
+### Patch Changes
+
+- Updated dependencies [cc50c9c]
+  - @llamaindex/env@0.1.28
+  - @llamaindex/core@0.5.1
+
+## 0.1.0
+
+### Minor Changes
+
+- 6a4a737: Remove re-exports from llamaindex main package
+
+### Patch Changes
+
+- Updated dependencies [6a4a737]
+- Updated dependencies [d924c63]
+  - @llamaindex/core@0.5.0
+
 ## 0.0.33
 
 ### Patch Changes