+++ packages/providers/vercel/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @llamaindex/vercel
 
+## 0.0.12
+
+### Patch Changes
+
+- @llamaindex/core@0.5.1
+
+## 0.0.11
+
+### Patch Changes
+
+- Updated dependencies [6a4a737]
+- Updated dependencies [d924c63]
+  - @llamaindex/core@0.5.0
+
 ## 0.0.10
 
 ### Patch Changes