+++ packages/providers/vllm/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @llamaindex/vllm
 
+## 0.0.23
+
+### Patch Changes
+
+- @llamaindex/openai@0.1.53
+
+## 0.0.22
+
+### Patch Changes
+
+- @llamaindex/openai@0.1.52
+
 ## 0.0.21
 
 ### Patch Changes